backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.12.0
qdrant-client==1.6.9
markdown==3.5.1
httpx==0.27.0
requests==2.31.0
psycopg[binary]>=3.1.0,<4.0.0